backend/requirements.txt
# FastAPI and ASGI server
fastapi
uvicorn[standard]
gunicorn

# Database
sqlalchemy
psycopg2-binary
alembic

# Authentication & Security
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart
PyJWT
argon2-cffi  # For Argon2 password hashing

# Pydantic
pydantic
pydantic-settings
email-validator

# AI Service
groq
requests

# PDF Generation
markdown
jinja2
weasyprint

# Storage
boto3
minio

# Background Tasks
celery
redis
rq

# Utils
python-dotenv
pillow
pypdf2

# Development & Testing
pytest
pytest-asyncio
pytest-cov
httpx
factory-boy
faker
pytest-mock
fakeredis

# Logging & Monitoring
structlog

# Security & Rate Limiting
slowapi>=0.1.9
bleach>=6.0.0
cryptography>=41.0.0
validators>=0.22.0